Confirm before playing music

+++ client/managers/youtube.py
@@ -4,6 +4,11 @@ from subprocess import Popen
 import requests
 from bs4 import BeautifulSoup
 
+import sys, os
+pardir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+sys.path.append(pardir)
+import utils.tts as tts
+from utils.speechrecognition import confirm
 import traceback
 
 class Manager:
@@ -16,14 +21,27 @@ class Manager:
     def execute(self, acts):
         for act in acts:
             if act["platform"] == "play_youtube":
+                if "confirmation" in act:
+                    tts.say(act['confirmation'])
+                    time.sleep(0.5)
+                    ans = confirm()
+                    if ans is None:
+                        tts.say("上手く聞こえませんでした")
+                        return
+                    elif not ans:
+                        tts.say("わかりました，再生をキャンセルします")
+                        return
+
                 try:
                     Popen('pkill -9 mpv', shell=True)
                     time.sleep(0.3)
+                    tts.say(act['data'] + "を検索します")
                     try:
                         youtube_result = Youtube(act['data'], result=1)
                         Popen("mpv '" + youtube_result.url[0] + "' --loop --no-video > /dev/null 2>&1", shell=True)
                     except:
-                        Popen("mpv https://www.youtube.com/watch?v=HKKe7p44PDY --loop --no-video > /dev/null 2>&1", shell=True)
+                        tts.say(act['data'] + "は見つかりませんでした")
+                        # Popen("mpv https://www.youtube.com/watch?v=HKKe7p44PDY --loop --no-video > /dev/null 2>&1", shell=True)
                     self.now_playing = act['data']
                 except:
                     traceback.print_exc()
@@ -44,33 +62,9 @@ class Youtube():
 
         self.data = [h3 for h3 in h3s]
         self.url = ["https://www.youtube.com" + h3.a.get('href') for h3 in h3s]
-        self.title = [h3.a.get("title") for h3 in h3s]
-        self.id = [h3.a.get("href").split("=")[-1] for h3 in h3s]
-        self.embed = ["https://www.youtube.com/embed/" + h3.a.get("href").split("=")[-1] for h3 in h3s]
-        self.time = [h3.span.text.replace(" - 長さ: ","").replace("。","") for h3 in h3s]
-        self.info = [h3.text for h3 in h3s] # >>タイトル　- 長さ：00:00。
-
-    def select(self):
-        values = {"url":self.url,"title":self.title,"id":self.id,"embed":self.embed,"time":self.time}
-        info = self.info
-        for i in range(len(info)):
-            print("%s:%s" % (i,info[i]))
-        while True:
-            try:
-                num = int(input("番号:"))
-                break
-            except:
-                print("番号を正しく入力してください。")
-        results = {
-            "url":values["url"][num],
-            "title":values["title"][num],
-            "id":values["id"][num],
-            "embed":values["embed"][num],
-            "time":values["time"][num],
-            }
-        return results
 
 if __name__ == "__main__":
-    youtubeplayer = Manager('sample', '1.0.0.0', None)
-    youtubeplayer.execute([{'platform': 'play_youtube', 'data': 'n'}])
+    # youtubeplayer = Manager('sample', '1.0.0.0', None)
+    # youtubeplayer.execute([{'platform': 'play_youtube', 'data': 'n'}])
     # youtubeplayer.execute([{'platform': 'stop_youtube', 'data': ''}])
+    yt = Youtube('nujabes')
